+++ src/components/Redux/store.js
@@ -1,5 +1,11 @@
-import { createStore } from 'redux';
-import reducer from './reducer';
+import { combineReducers, createStore } from 'redux';
+import AddRemove from './reducers/addRemove';
+import inDeCrement from './reducers/inDeCrement';
+
+const reducer = combineReducers({
+  AddRemove,
+  inDeCrement,
+});
 
 const store = createStore(reducer);
